l4re_vfs: Add implementation of pwritev.
Adds a pwritev implementation to Dataspace-backed files. This allows to
change the contents of files accessed via the VFS. In order to write a
file, an application must have write permission for the respective
Dataspace capability. Note that for now, write operations beyond EOF,
i.e., write operations that would require an extension of a file, are
not supported.

Co-authored-by: Viktor Reusch <viktor.reusch@barkhauseninstitut.org>

Change-Id: I38d5efb0596d3601a3e29d4d615152541e372d87
Signed-off-by: Till Miemietz <till.miemietz@barkhauseninstitut.org>